+++ sdk/python/src/muxai/__init__.py
@@ -1,1 +1,23 @@
 """muxai Python SDK."""
+
+from .client import Client, ClientConfig
+from .errors import ErrorCode, MuxaiError
+from .provider import Provider
+from .providers import ClaudeProvider, CursorProvider, VibeProvider
+from .types import Message, ProviderName, Request, Response, Role
+
+__all__ = [
+    "Client",
+    "ClientConfig",
+    "ClaudeProvider",
+    "CursorProvider",
+    "ErrorCode",
+    "Message",
+    "MuxaiError",
+    "Provider",
+    "ProviderName",
+    "Request",
+    "Response",
+    "Role",
+    "VibeProvider",
+]